pyproject.toml
[project]
name = "sales-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.7",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "firecrawl>=2.16.5",
    "firecrawl-py>=2.16.5",
    "html2text>=2025.4.15",
    "langchain>=0.3.27",
    "openai>=1.99.1",
    "packaging>=25.0",
    "protobuf==5.29.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scrapy>=2.13.3",
    "sqlalchemy>=2.0.42",
    "tavily-python>=0.7.10",
    "weaviate-client>=4.16.6",
]
